apidoc compatible documentation
